# swagger/crowd-lending.yml
swagger: '2.0'
info:
  version: '1.0'
  title: Crowd Lending
  description: API Suite for Crowd Lending platforms
host: 'localhost:8080'
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Authorization:
    name: Authorization
    type: apiKey
    in: header
  X-API-KEY:
    name: X-API-KEY
    type: apiKey
    in: header
paths:
  '/loans/{loan_id}':
    parameters:
      - name: loan_id
        in: path
        required: true
        type: string
    get:
      operationId: get-loan
      summary: Get Loan
      produces:
        - application/json
      tags:
        - Loans
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/loan-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    put:
      operationId: update-loan
      summary: Update Loan
      tags:
        - Loans
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/loan'
      responses:
        '200':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/loan-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    delete:
      operationId: delete-loan
      summary: Delete Loan
      tags:
        - Loans
      responses:
        '204':
          description: 'Successful Response'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
  /loans:
    get:
      operationId: list-loans
      summary: List Loans
      tags:
        - Loans
      parameters:
        - name: loan_type
          in: query
          description: Investor Id of the investments
          type: string
        - name: from_date
          in: query
          description: From date to get all the investments created after this date
          type: string
          format: date
        - name: to_date
          in: query
          description: to date to get all the investments created before this date. Default is current date
          type: string
          format: date
      responses:
        '200':
          description: 'Successful Response'
          schema:
            type: array
            items:
              $ref: '#/definitions/loan-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    post:
      operationId: create-loan
      summary: Create Loan
      tags:
        - Loans
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/loan'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/loan-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
  '/loans/{loan_id}/investments':
    parameters:
      - name: loan_id
        in: path
        required: true
        type: string
    post:
      operationId: Create-Investment
      summary: Investment to the loan
      produces:
        - application/json
      tags:
        - Loans
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/investment'
      responses:
        '201':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/investment-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    get:
      operationId: get-Investments
      summary: Get Investments of a loan
      produces:
        - application/json
      tags:
        - Loans
      parameters:
        - name: investor_id
          in: query
          description: Investor Id of the investments
          type: string
        - name: from_date
          in: query
          description: From date to get all the investments for a loan
          type: string
          format: date
        - name: to_date
          in: query
          description: to date to get all the investments for a loan. Default is current date
          type: string
          format: date
      responses:
        '200':
          description: 'Successful Response'
          schema:
            type: array
            items:
              $ref: '#/definitions/investment-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
  '/loans/{loan_id}/repayments':
    parameters:
      - name: loan_id
        in: path
        required: true
        type: string
    post:
      operationId: Create-Repayment
      summary: Repayment of the loan
      tags:
        - Loans
      parameters:
        - name: body
          in: body
          schema:
            $ref: '#/definitions/repayment'
      responses:
        '201':
          description: 'Successful Response'
          schema:
            $ref: '#/definitions/repayment-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
    get:
      operationId: get-Repayments
      summary: Get all the Repayment details of the loan
      produces:
        - application/json
      tags:
        - Loans
      parameters:
        - name: from_date
          in: query
          description: From date to get all the repayments for a loan
          type: string
          format: date
        - name: to_date
          in: query
          description: to date to get all the repayments for a loan. Default is current date
          type: string
          format: date
      responses:
        '200':
          description: 'Successful Response'
          schema:
            type: array
            items:
              $ref: '#/definitions/repayment-info'
        '400':
          description: 'Bad Request'
          schema:
            $ref: '#/definitions/generic-error'
        '401':
          description: 'Authentication Error'
          schema:
            $ref: '#/definitions/generic-error'
        '403':
          description: 'Authorization Failed'
          schema:
            $ref: '#/definitions/generic-error'
        '404':
          description: 'Resource Not Found'
          schema:
            $ref: '#/definitions/generic-error'
        '500':
          description: 'Internal Server Error'
          schema:
            $ref: '#/definitions/server-error'
      security:
        - Authorization: []
        - X-API-KEY: []
...
definitions:
...
  loan:
    title: Loan
    type: object
    properties:
      description:
        type: string
      amount:
        type: number
      currency:
        type: string
      borrower_id:
        type: string
      term:
        type: string
      interest_rate:
        type: string
      loan_type:
        type: string
        enum:
          - PERSONAL
          - BUSINESS
      status:
        type: string
      created_date:
        type: string
        format: date
      repayment_plan:
        type: string
      repayment_start_date:
        type: string
        format: date
      projected_repayment_end_date:
        type: string
        format: date
      repaid_amount:
        type: number
      loan_account:
        $ref: '#/definitions/account-details'
  server-error:
    title: Server Error
    type: object
    properties:
      code:
        type: string
      message:
        type: string
      tracking_id:
        type: string
  generic-error:
    title: Generic Error
    type: object
    properties:
      code:
        type: string
      message:
        type: string
  loan-info:
    title: Loan Info with ID
    type: object
    properties:
      id:
        type: string
      loan:
        $ref: '#/definitions/loan'
      investors:
        type: array
        items:
          type: string
          description: List of Investor Id's
...
  investment:
    title: Investment details
    type: object
    properties:
      investor_id:
        type: string
      amount:
        type: number
      currency:
        type: string
      payment_transaction_id:
        type: string
      payment_date:
        type: string
        format: date
      comments:
        type: string
  investment-info:
    title: Investment Info
    type: object
    properties:
      id:
        type: string
      loan_id:
        type: string
      investment:
        $ref: '#/definitions/investment'
  repayment:
    title: Repayment details
    type: object
    properties:
      borrower_id:
        type: string
      amount:
        type: number
      currency:
        type: string
      payment_transaction_id:
        type: string
      payment_date:
        type: string
        format: date
      comments:
        type: string
  repayment-info:
    title: Repayment Info
    type: object
    properties:
      id:
        type: string
      loan_id:
        type: string
      repayment:
        $ref: '#/definitions/repayment'
  account-details:
    title: Account Details
    type: object
    properties:
      bank_name:
        type: string
      bank_swift_code:
        type: string
      account_number:
        type: string
      account_name:
        type: string